move more condition logic into lvl_script_conditions
