+++ Makefile
@@ -246,7 +246,6 @@ obj/lvl_filesdk1.o \
 obj/lvl_script.o \
 obj/lvl_script_commands.o \
 obj/lvl_script_lib.o \
-obj/lvl_script_statehandler.o \
 obj/lvl_script_conditions.o \
 obj/lvl_script_parser.o \
 obj/magic.o \
